Impl dos repositories

+++ src/main/java/model/Categoria.java
@@ -18,7 +18,7 @@ public class Categoria {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int codigo;
+	private Long codigo;
 
 	@Column(length = 100)
 	private String nome;
